pyproject.toml
[project]
name = "google-cloud-datastore-odm"
version = "0.1.0"
description = "Minimal ORM for Google Cloud Datastore inspired by NDB"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-datastore>=2.23.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-env>=1.2.0",
    "ruff>=0.14.10",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
env = [
    "GOOGLE_CLOUD_PROJECT=google-cloud-datastore-odm-test",
    "DATASTORE_EMULATOR_HOST=localhost:20001"
]
log_format="%(asctime)s %(levelname)s %(message)s"
log_level="INFO"
log_date_format="%Y-%m-%d %H:%M:%S"
log_cli="true"
log_cli_format="%(asctime)s %(levelname)s %(message)s"
log_cli_date_format="%Y-%m-%d %H:%M:%S"
addopts="-p no:warnings"
